feat: Company search 구현 - user, company 도메인 분리로 인한 패키지 수정
