+++ auth-service/src/main/java/org/mosaic/auth/company/application/dtos/UpdateCompanyDto.java
@@ -1,0 +1,27 @@
+package org.mosaic.auth.company.application.dtos;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import org.mosaic.auth.company.domain.entity.company.CompanyType;
+
+@Getter
+@Builder(access = AccessLevel.PRIVATE)
+public class UpdateCompanyDto {
+
+  private Long companyId;
+  private String name;
+  private String address;
+  private CompanyType companyType;
+
+  public static UpdateCompanyDto create(Long id, String name, String address,
+      CompanyType companyType) {
+    return UpdateCompanyDto.builder()
+        .companyId(id)
+        .name(name)
+        .address(address)
+        .companyType(companyType)
+        .build();
+  }
+
+}
