+++ auth-service/src/main/java/org/mosaic/auth/company/application/dtos/UpdateCompanyHubIdDto.java
@@ -1,0 +1,21 @@
+package org.mosaic.auth.company.application.dtos;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder(access = AccessLevel.PRIVATE)
+public class UpdateCompanyHubIdDto {
+
+  private Long companyId;
+  private Long hubId;
+
+  public static UpdateCompanyHubIdDto create(Long companyId, Long hubId) {
+    return UpdateCompanyHubIdDto.builder()
+        .companyId(companyId)
+        .hubId(hubId)
+        .build();
+  }
+
+}
