+++ auth-service/src/main/java/org/mosaic/auth/company/domain/entity/company/CompanyType.java
@@ -1,0 +1,5 @@
+package org.mosaic.auth.company.domain.entity.company;
+
+public enum CompanyType {
+    SUPPLY, DEMAND
+}
