--- auth-service/src/main/java/org/mosaic/auth/config/QueryDslConfig.java
@@ -1,20 +1,0 @@
-package org.mosaic.auth.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@RequiredArgsConstructor
-public class QueryDslConfig {
-
-    private final EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory jPAQueryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-
-}
