--- auth-service/src/main/java/org/mosaic/auth/domain/entity/company/BaseEntity.java
@@ -1,62 +1,0 @@
-package org.mosaic.auth.domain.entity.company;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public class BaseEntity {
-    @CreatedDate
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    @CreatedBy
-    @Column(name = "created_by", updatable = false)
-    private Long createdBy;
-
-    @LastModifiedDate
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    @LastModifiedBy
-    @Column(name = "updated_by")
-    private Long updatedBy;
-
-    @Column(name = "deleted_at")
-    private LocalDateTime deletedAt;
-
-    @Column(name = "deleted_by")
-    private Long deletedBy;
-
-    @Column(name = "is_deleted")
-    private Boolean isDeleted = false;
-
-    @Column(name = "is_public")
-    private Boolean isPublic = true;
-
-
-    public void softDelete() {
-        this.deletedAt = LocalDateTime.now();
-        this.isDeleted = true;
-    }
-
-    public void delete() {
-        this.deletedAt = LocalDateTime.now();
-        this.isDeleted = true;
-        this.isPublic = false;
-    }
-
-    public void updatePrivate() {
-        this.isPublic = false;
-    }
-}
