--- auth-service/src/main/java/org/mosaic/auth/domain/entity/company/CompanyType.java
@@ -1,5 +1,0 @@
-package org.mosaic.auth.domain.entity.company;
-
-public enum CompanyType {
-    SUPPLY, DEMAND
-}
