--- auth-service/src/main/java/org/mosaic/auth/domain/entity/user/UserRole.java
@@ -1,5 +1,0 @@
-package org.mosaic.auth.domain.entity.user;
-
-public enum UserRole {
-    COMPANY, DELIVERY, HUB_MANAGER, MASTER
-}
