+++ auth-service/src/main/java/org/mosaic/auth/user/domain/entity/user/UserRole.java
@@ -1,0 +1,5 @@
+package org.mosaic.auth.user.domain.entity.user;
+
+public enum UserRole {
+    COMPANY, DELIVERY, HUB_MANAGER, MASTER
+}
