+++ auth-service/src/main/java/org/mosaic/auth/user/infrastructure/UserRepositoryImpl.java
@@ -1,0 +1,11 @@
+package org.mosaic.auth.user.infrastructure;
+
+import org.mosaic.auth.user.domain.entity.user.User;
+import org.mosaic.auth.user.domain.repository.UserRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepositoryImpl extends JpaRepository<User, Long>,
+    UserRepository
+{
+
+}
